requirements.txt
langchain-core==0.3.75
langchain-community==0.3.29
faiss-cpu==1.12.0
accelerate==1.10.1
pypdf==6.0.0
flashrank==0.2.10
chromadb==1.0.20
groq==0.31.0
protobuf>=3.20.3,<5
google-generativeai==0.8.5
nltk==3.9.1
python-dotenv==1.1.1
google-ai-generativelanguage==0.6.18